refactor(users): split lib/userSS into lib/users/svcSS + lib/auth/svcSS + lib/auth/types + lib/users/types
AuthTypeMetadata → lib/auth/types
Auth SS functions → lib/auth/svcSS
User SS functions → lib/users/svcSS
CustomRefreshTokenResponse → lib/users/types
