+++ web/src/app/auth/verify-email/page.tsx
@@ -1,8 +1,6 @@
-import {
-  AuthTypeMetadata,
-  getAuthTypeMetadataSS,
-  getCurrentUserSS,
-} from "@/lib/userSS";
+import { getCurrentUserSS } from "@/lib/users/svcSS";
+import { getAuthTypeMetadataSS } from "@/lib/auth/svcSS";
+import { AuthTypeMetadata } from "@/lib/auth/types";
 import Verify from "./Verify";
 import { User } from "@/lib/types";
 import { redirect } from "next/navigation";
