+++ web/src/app/auth/waiting-on-verification/page.tsx
@@ -1,8 +1,6 @@
-import {
-  AuthTypeMetadata,
-  getAuthTypeMetadataSS,
-  getCurrentUserSS,
-} from "@/lib/userSS";
+import { getCurrentUserSS } from "@/lib/users/svcSS";
+import { getAuthTypeMetadataSS } from "@/lib/auth/svcSS";
+import { AuthTypeMetadata } from "@/lib/auth/types";
 import { redirect } from "next/navigation";
 import { User } from "@/lib/types";
 import { RequestNewVerificationEmail } from "./RequestNewVerificationEmail";
