+++ web/src/lib/auth/requireAuth.ts
@@ -1,9 +1,7 @@
 import { User, UserRole } from "@/lib/types";
-import {
-  AuthTypeMetadata,
-  getAuthTypeMetadataSS,
-  getCurrentUserSS,
-} from "@/lib/userSS";
+import { getCurrentUserSS } from "@/lib/users/svcSS";
+import { getAuthTypeMetadataSS } from "@/lib/auth/svcSS";
+import { AuthTypeMetadata } from "@/lib/auth/types";
 import { AuthType } from "@/lib/constants";
 
 /**
