+++ web/src/lib/auth/types.ts
@@ -1,0 +1,11 @@
+import { AuthType } from "@/lib/constants";
+
+export interface AuthTypeMetadata {
+  authType: AuthType;
+  autoRedirect: boolean;
+  requiresVerification: boolean;
+  anonymousUserEnabled: boolean | null;
+  passwordMinLength: number;
+  hasUsers: boolean;
+  oauthEnabled: boolean;
+}
