--- web/src/lib/userSS.ts
@@ -1,200 +1,0 @@
-import { cookies } from "next/headers";
-import { User } from "./types";
-import { buildUrl, UrlBuilder } from "./utilsSS";
-import { ReadonlyRequestCookies } from "next/dist/server/web/spec-extension/adapters/request-cookies";
-import {
-  AuthType,
-  NEXT_PUBLIC_CLOUD_ENABLED,
-  SERVER_SIDE_ONLY__AUTH_COOKIE_NAME,
-} from "./constants";
-
-export interface AuthTypeMetadata {
-  authType: AuthType;
-  autoRedirect: boolean;
-  requiresVerification: boolean;
-  anonymousUserEnabled: boolean | null;
-  passwordMinLength: number;
-  hasUsers: boolean;
-  oauthEnabled: boolean;
-}
-
-export const getAuthTypeMetadataSS = async (): Promise<AuthTypeMetadata> => {
-  const res = await fetch(buildUrl("/auth/type"));
-  if (!res.ok) {
-    throw new Error("Failed to fetch data");
-  }
-
-  const data: {
-    auth_type: string;
-    requires_verification: boolean;
-    anonymous_user_enabled: boolean | null;
-    password_min_length: number;
-    has_users: boolean;
-    oauth_enabled: boolean;
-  } = await res.json();
-
-  let authType: AuthType;
-
-  // Override fastapi users auth so we can use both
-  if (NEXT_PUBLIC_CLOUD_ENABLED) {
-    authType = AuthType.CLOUD;
-  } else {
-    authType = data.auth_type as AuthType;
-  }
-
-  // for SAML / OIDC, we auto-redirect the user to the IdP when the user visits
-  // Onyx in an un-authenticated state
-  if (authType === AuthType.OIDC || authType === AuthType.SAML) {
-    return {
-      authType,
-      autoRedirect: true,
-      requiresVerification: data.requires_verification,
-      anonymousUserEnabled: data.anonymous_user_enabled,
-      passwordMinLength: data.password_min_length,
-      hasUsers: data.has_users,
-      oauthEnabled: data.oauth_enabled,
-    };
-  }
-  return {
-    authType,
-    autoRedirect: false,
-    requiresVerification: data.requires_verification,
-    anonymousUserEnabled: data.anonymous_user_enabled,
-    passwordMinLength: data.password_min_length,
-    hasUsers: data.has_users,
-    oauthEnabled: data.oauth_enabled,
-  };
-};
-
-const getOIDCAuthUrlSS = async (nextUrl: string | null): Promise<string> => {
-  const url = UrlBuilder.fromClientUrl("/api/auth/oidc/authorize");
-  if (nextUrl) {
-    url.addParam("next", nextUrl);
-  }
-  url.addParam("redirect", true);
-
-  return url.toString();
-};
-
-const getGoogleOAuthUrlSS = async (nextUrl: string | null): Promise<string> => {
-  const url = UrlBuilder.fromClientUrl("/api/auth/oauth/authorize");
-  if (nextUrl) {
-    url.addParam("next", nextUrl);
-  }
-  url.addParam("redirect", true);
-
-  return url.toString();
-};
-
-const getSAMLAuthUrlSS = async (nextUrl: string | null): Promise<string> => {
-  const url = UrlBuilder.fromInternalUrl("/auth/saml/authorize");
-  if (nextUrl) {
-    url.addParam("next", nextUrl);
-  }
-
-  const res = await fetch(url.toString());
-  if (!res.ok) {
-    throw new Error("Failed to fetch data");
-  }
-
-  const data: { authorization_url: string } = await res.json();
-  return data.authorization_url;
-};
-
-export const getAuthUrlSS = async (
-  authType: AuthType,
-  nextUrl: string | null
-): Promise<string> => {
-  // Returns the auth url for the given auth type
-
-  switch (authType) {
-    case AuthType.BASIC:
-      return "";
-    case AuthType.GOOGLE_OAUTH: {
-      return await getGoogleOAuthUrlSS(nextUrl);
-    }
-    case AuthType.CLOUD: {
-      return await getGoogleOAuthUrlSS(nextUrl);
-    }
-    case AuthType.SAML: {
-      return await getSAMLAuthUrlSS(nextUrl);
-    }
-    case AuthType.OIDC: {
-      return await getOIDCAuthUrlSS(nextUrl);
-    }
-  }
-};
-
-const logoutStandardSS = async (headers: Headers): Promise<Response> => {
-  return await fetch(buildUrl("/auth/logout"), {
-    method: "POST",
-    headers: headers,
-  });
-};
-
-const logoutSAMLSS = async (headers: Headers): Promise<Response> => {
-  return await fetch(buildUrl("/auth/saml/logout"), {
-    method: "POST",
-    headers: headers,
-  });
-};
-
-export const logoutSS = async (
-  authType: AuthType,
-  headers: Headers
-): Promise<Response | null> => {
-  switch (authType) {
-    case AuthType.SAML: {
-      return await logoutSAMLSS(headers);
-    }
-    default: {
-      return await logoutStandardSS(headers);
-    }
-  }
-};
-
-export const getCurrentUserSS = async (): Promise<User | null> => {
-  try {
-    const cookieString = processCookies(await cookies());
-
-    const response = await fetch(buildUrl("/me"), {
-      credentials: "include",
-      next: { revalidate: 0 },
-      headers: {
-        cookie: cookieString,
-      },
-    });
-
-    if (!response.ok) {
-      return null;
-    }
-
-    const user = await response.json();
-    return user;
-  } catch (e) {
-    console.log(`Error fetching user: ${e}`);
-    return null;
-  }
-};
-
-export const processCookies = (cookies: ReadonlyRequestCookies): string => {
-  let cookieString = cookies
-    .getAll()
-    .map((cookie) => `${cookie.name}=${cookie.value}`)
-    .join("; ");
-
-  // Inject debug auth cookie for local development against remote backend (only if not already present)
-  if (process.env.DEBUG_AUTH_COOKIE && process.env.NODE_ENV === "development") {
-    const hasAuthCookie = cookieString
-      .split(/;\s*/)
-      .some((c) => c.startsWith(`${SERVER_SIDE_ONLY__AUTH_COOKIE_NAME}=`));
-    if (!hasAuthCookie) {
-      const debugCookie = `${SERVER_SIDE_ONLY__AUTH_COOKIE_NAME}=${process.env.DEBUG_AUTH_COOKIE}`;
-      cookieString = cookieString
-        ? `${cookieString}; ${debugCookie}`
-        : debugCookie;
-    }
-  }
-
-  return cookieString;
-};
