+++ web/src/lib/utilsSS.ts
@@ -1,6 +1,6 @@
 import { cookies } from "next/headers";
 import { HOST_URL, INTERNAL_URL } from "./constants";
-import { processCookies } from "@/lib/userSS";
+import { processCookies } from "@/lib/users/svcSS";
 
 export function buildClientUrl(path: string) {
   if (path.startsWith("/")) {
